[BUILD] Move EclipseTestThread into client.util package
