+++ stf.test/test/saros/stf/test/consistency/AddMultipleFilesTest.java
@@ -7,8 +7,8 @@ import static saros.stf.client.tester.SarosTester.BOB;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
+import saros.stf.client.util.EclipseTestThread;
 import saros.stf.client.util.Util;
-import saros.test.util.EclipseTestThread;
 
 public class AddMultipleFilesTest extends StfTestCase {
