+++ stf.test/test/saros/stf/test/consistency/EditDuringInvitationStressTest.java
@@ -13,10 +13,10 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.client.tester.SarosTester;
+import saros.stf.client.util.EclipseTestThread;
 import saros.stf.client.util.Util;
 import saros.stf.shared.Constants.TypeOfCreateProject;
 import saros.stf.test.stf.Constants;
-import saros.test.util.EclipseTestThread;
 
 public class EditDuringInvitationStressTest extends StfTestCase {
