+++ stf.test/test/saros/stf/test/editing/ConcurrentEditingInsert100CharactersTest.java
@@ -9,9 +9,9 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.client.tester.AbstractTester;
+import saros.stf.client.util.EclipseTestThread;
 import saros.stf.client.util.Util;
 import saros.stf.shared.Constants.TypeOfCreateProject;
-import saros.test.util.EclipseTestThread;
 
 public class ConcurrentEditingInsert100CharactersTest extends StfTestCase {
